Implementing Dataview - Refactoring selectors for InputsView and OutputsView
